Arrow function fix
For most arrow functions, the extension failed to output documentation.  This is because in JS most of them were being classified as VariableDeclarationLists or VariableStatements.  Inside of those types’ declarations was the Arrow Function that was not being grabbed initially.

So, those kinds were added to the kinds array in `src/utilities.ts`, and a handler for them was added, `_emitArrowFunction`, which outputs the name and then hands the ArrowFunction node over to `_emitFunctionExpression`.