Removed VariableDeclarationList function target
When attempting to determine the first parent of a funciton declaration, a kind of VariableDeclarationList was listed.  This would not work in the case where the function had modifiers like “export” and would overwrite them.  The VariableStatement takes modifiers into account and seems to always be the parent of a VariableDeclarationList.