Formatted returns for docs-core
The types would have been output if the code could have sensed them, but needed to be in `{}`.  This is now done and a `-` character is added to the end for a returns description.  If includeTypes is set, and type cannot be found or infered, `{any}` will be included for `@returns`.